+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Doesn't allow a property to have a default=true if enabled=false (todo #160)
+    - Replace 'allowWarn' and 'allowVerbose' just with 'warn' and 'verbose' (todo #161)
 
 ### 4.17.2